// src/user.d.ts
export { User }

interface User {
    name: string;
    surname: string;
    company:  string;
    email: string;
    phone: string;
    country: string;
    city: string;
    region: string;
    adress: string;
    index: string;
    paySystem: string;
    cardNumber1: string;
    cardNumber2: string;
    cardNumber3: string;
    cardNumber4: string;
}